Changes were on wrong gui file, corrected so there's no confusion.
